refactor: Update import paths for Header component in ExperimentCollectionDashboard and Header files
